% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidneyoutcomesintx.R
\name{kidneyoutcomesintx}
\alias{kidneyoutcomesintx}
\title{kidneyoutcomesintx}
\usage{
kidneyoutcomesintx(inputdataframe, outputdirectory)
}
\arguments{
\item{inputdataframe}{Data frame. Name of loaded dataframe containing the variales pt_id,    date_trans1,    date_lab,    sex,age_at_lab,creatinine_mgdl. See README for details.}

\item{outputdirectory}{Character. The path to the output directory where results will be saved.}
}
\value{
Individual dataframes containing:
\itemize{
\item Summary number of AKI events occurring in year 1 after transplant (or some defined baseline) for each patient ID
\item Summary of every AKI event that occurred for the individual, containing the start and end time in days from baseline and the length of the event. If given a value of 0 there were no AKI events for that individual
\item Summary of AKI events over all follow up, the number of which were >7 days and the identification of CKD progression and timing of progression in days with respect to baseline. (eGFR thresholds of 15, 30, 45, and 60)
}

and finally a summary table in a word document for serum creatinine measurement rates and frequencies.
}
\description{
Processes serum creatinine data to identify Acute Kidney Injury (AKI) events
and progression to Chronic Kidney Disease (CKD) stages within a defined
baseline context. Developed for use in islet transplantation but adaptable
for other settings.
}
\details{
The function requires an input file containing specific variables:
\itemize{
\item \code{pt_id}: Unique patient identifier.
\item \code{date_trans1}: Date of the first transplant/study entry, used as the "baseline" date.
\item \code{date_lab}: Date of the serum creatinine laboratory result.
\item \code{sex}: Define 1 for male, 2 for female.
\item \code{age_at_lab}: Age of the patient at the laboratory date (in years).
\item \code{creatinine_mgdl}: Serum creatinine levels (Units MUST BE mg/dL).
}
}
\examples{
\dontrun{
kidneyoutcomesintx("path/to/input.csv", "path/to/output/")
}

}
\author{
Alice Carr
}
\keyword{AKI}
\keyword{CKD}
\keyword{kidney}
\keyword{outcomes}
\keyword{transplant}
